% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gquote.R
\name{getIntradayPrice}
\alias{getIntradayPrice}
\title{Download intraday data from Google Finance or Yahoo Finance}
\usage{
getIntradayPrice(ticker, src = "google", period = 1, interval = 5,
  tz = NULL, auto.assign = TRUE, time.shift = 0)
}
\arguments{
\item{ticker}{ticker for the data to download, include source (exchange)}

\item{src}{source, can be either "google" or "yahoo"}

\item{period}{time period in integer # of days}

\item{interval}{time interval in integer # of minutes}

\item{tz}{timezone, set to NULL by default}

\item{auto.assign}{similar to getSymbols in quantmod, set to false by default}

\item{time.shift}{time shift to apply (in hours) to align the time stamp}
}
\value{
an xts time series
}
\description{
Download intraday data from Google Finance or Yahoo Finance in xts format
}
\examples{
x <- gquote::getIntradayPrice("INDEXSP:.INX",period=5, interval=5)
tail(x)
}

